# src/internal/api/docs/proto/main.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "proto/main.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "BecomeBetter"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/main_categories": {
      "get": {
        "operationId": "BecomeBetter_MainCategories",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/exampleMainCategoriesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "BecomeBetter"
        ]
      },
      "post": {
        "operationId": "BecomeBetter_AddCategories",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/exampleMainCategoriesMessage"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "description",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "mainCategory",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "BecomeBetter"
        ]
      }
    }
  },
  "definitions": {
    "exampleMainCategoriesMessage": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "mainCategory": {
          "type": "string"
        }
      }
    },
    "exampleMainCategoriesResponse": {
      "type": "object",
      "properties": {
        "MainCategories": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/exampleMainCategoriesMessage"
          }
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}